Added Form to choose values
to Fulfill smynos's request.

Most of the "Bulk Edits" buttons can now be Ctrl+clicked, which displays
a small form in which you can select the value to apply for that cheat
(instead of maximum).

Also removed the "blacked stage sprites" feature for now, I don't like
it anymore.

Also fixed the "SetEscalationStep" function : now it makes the passed
value the next step to beat instead of the last one beaten (giving 50 as
an arg would make you beat step 51 for example).