Edit Pokathlon (better)
Added support for enabling the "paused" state that allow you to resume
your Pokathlon session.

Now "Edit Pokathlon" button sets step to 50, moves to 99 , opponent as
Mega Mewtwo Y and also activates the "paused" mode so that this session
can be resumed next time the game is launched.

Inverted "Random" button behaviour : now click it to get a random legit
opponent, or ctrl+click it to get a random opponent among the current
400 stages.