Added verbose Skill output
Available skills for a Pokémonare now written in full text in a
selectable list.
Hover over that list for skill description.
This adds one resource file, "message_Pokedex_US.bin", from which the
strings are taken. It is found in Message_US folder.
English only for now, won't work with another region's strings.
Updated resource files to v1.3.17 of the game.